# servers/center/src/main/java/com/wondercare/center/api/CenterApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.wondercare.center.api;

import com.wondercare.center.api.model.Center;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-01-19T21:41:34.831+08:00")

@Api(value = "center", description = "the center API")
public interface CenterApi {

    @ApiOperation(value = "Find center by ID", nickname = "getCenterById", notes = "Returns a single center", response = Center.class, tags={ "center", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Center.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Center not found") })
    @RequestMapping(value = "/center/{centerId}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Center> getCenterById(@ApiParam(value = "ID of center to return", required = true) @PathVariable("centerId") Long centerId);

}