# servers/center/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: "Wonder Center API"
  version: "1.0.0"
  title: "Wondercare Center API"
  termsOfService: ""
  contact:
    email: "hutingung@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "wondercare.my"
basePath: "/v1"
tags:
- name: "center"
  description: "Everything about your center"
  externalDocs:
    description: "Find out more"
    url: "http://wondercare.my"
schemes:
- "https"
- "http"
paths:
  /center/{centerId}:
    get:
      tags:
      - "center"
      summary: "Find center by ID"
      description: "Returns a single center"
      operationId: "getCenterById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "centerId"
        in: "path"
        description: "ID of center to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Center"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Center not found"
definitions:
  Center:
    type: "object"
    required:
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        example: "kids cove"
      businessId:
        type: "string"
      address:
        $ref: "#/definitions/Address"
      location:
        $ref: "#/definitions/Location"
      contact:
        $ref: "#/definitions/Contact"
      type:
        $ref: "#/definitions/CenterType"
  Address:
    type: "object"
    properties:
      address:
        type: "string"
      postcode:
        type: "integer"
      state:
        type: "string"
      countryCode:
        type: "string"
  Contact:
    type: "object"
    properties:
      email:
        type: "string"
      telNo:
        type: "string"
  Location:
    type: "object"
    properties:
      longitude:
        type: "integer"
      latitude:
        type: "integer"
  CenterType:
    type: "string"
    enum: 
      - INSTITUITION
      - HOME
      - UNKNOWN
      - WORKPLACE
    
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"